Remove duplicated DynamoDB state store tests
